requirements.txt
langchain==0.2.1
langchain-core==0.2.3
langchain-community
langchain-groq
fastapi
uvicorn
python-dotenv
duckduckgo-search
apscheduler